grant_agent/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
gunicorn>=21.2.0
httpx>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.11
playwright>=1.44.0
apscheduler>=3.10.4
anthropic>=0.28.0
openai>=1.30.0
python-dotenv>=1.0.0
lxml>=5.2.0
aiosqlite>=0.20.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
